# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Venice AI Legal Analysis Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /payments:
    get:
      operationId: getPayments
      tags: [payments]
      summary: Get crypto payment logs
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Payment log entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PaymentEntry"
...
  /payments/charge:
    post:
      operationId: createCharge
      tags: [payments]
      summary: Create a USDC charge request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateChargeInput"
      responses:
        "201":
          description: Charge created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChargeRequest"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /payments/charge/{id}:
    get:
      operationId: getCharge
      tags: [payments]
      summary: Get charge details for payment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Charge details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChargeDetail"
        "404":
          description: Charge not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteCharge
      tags: [payments]
      summary: Cancel/expire a pending charge
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Charge cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Cannot delete paid charge
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Charge not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /payments/confirm:
    post:
      operationId: confirmPayment
      tags: [payments]
      summary: Confirm a USDC payment with transaction hash
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmPaymentInput"
      responses:
        "200":
          description: Payment confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentEntry"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Transaction already recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /payments/delegation:
    get:
      operationId: getDelegation
      tags: [payments]
      summary: Get current delegation status and EIP-712 type info
      responses:
        "200":
          description: Delegation status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DelegationInfo"
    post:
      operationId: submitDelegation
      tags: [payments]
      summary: Submit a signed EIP-712 delegation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitDelegationInput"
      responses:
        "200":
          description: Delegation stored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DelegationStatus"
        "400":
          description: Invalid delegation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /payments/identity/register:
    post:
      operationId: registerAgentIdentity
      tags: [payments]
      summary: Register agent on ERC-8004 IdentityRegistry
      responses:
        "200":
          description: Registration result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegistrationResult"
        "400":
          description: Registration failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /payments/swap:
    post:
      operationId: executeSwap
      tags: [payments]
      summary: Manually trigger a USDC to ETH swap via Uniswap
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SwapInput"
      responses:
        "200":
          description: Swap executed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SwapResult"
        "400":
          description: Swap failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Delegation denied or unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Uniswap not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    PaymentEntry:
      type: object
      properties:
        id:
          type: string
        txHash:
          type: string
        from:
          type: string
        to:
          type: string
        amount:
          type: string
        token:
          type: string
        status:
          type: string
          enum: [pending, confirmed, failed]
        timestamp:
          type: string
          format: date-time
        network:
          type: string
        paymentMethod:
          type: string
          enum: [direct, locus, swap]
      required:
        - id
        - amount
        - token
        - status
        - timestamp
...
    ChargeRequest:
      type: object
      properties:
        id:
          type: string
        amount:
          type: string
        label:
          type: string
        status:
          type: string
          enum: [pending, paid, expired]
        createdAt:
          type: string
          format: date-time
        paidAt:
          type: string
          format: date-time
        txHash:
          type: string
        paidFrom:
          type: string
        locusWalletAddress:
          type: string
      required:
        - id
        - amount
        - status
        - createdAt
    CreateChargeInput:
      type: object
      properties:
        amount:
          type: number
        label:
          type: string
      required:
        - amount
    ChargeDetail:
      type: object
      properties:
        id:
          type: string
        amount:
          type: string
        label:
          type: string
        status:
          type: string
          enum: [pending, paid, expired]
        createdAt:
          type: string
          format: date-time
        paidAt:
          type: string
          format: date-time
        txHash:
          type: string
        paidFrom:
          type: string
        walletAddress:
          type: string
        usdcContract:
          type: string
        network:
          type: string
        chainId:
          type: integer
        paymentMethod:
          type: string
          enum: [direct, locus]
        locusWalletAddress:
          type: string
      required:
        - id
        - amount
        - status
        - createdAt
        - walletAddress
        - network
        - chainId
    ConfirmPaymentInput:
      type: object
      properties:
        txHash:
          type: string
          description: Transaction hash to verify on-chain
        chargeId:
          type: string
          description: Optional charge ID to link this payment to
      required:
        - txHash
    DelegationStatus:
      type: object
      properties:
        active:
          type: boolean
        delegator:
          type: string
        delegate:
          type: string
        dailyLimitUsdc:
          type: number
        dailyUsedUsdc:
          type: number
        dailyRemainingUsdc:
          type: number
        expiresAt:
          type: string
          format: date-time
        expired:
          type: boolean
        signedAt:
          type: string
          format: date-time
        reason:
          type: string
      required:
        - active
    DelegationInfo:
      type: object
      properties:
        active:
          type: boolean
        delegator:
          type: string
        delegate:
          type: string
        dailyLimitUsdc:
          type: number
        dailyUsedUsdc:
          type: number
        dailyRemainingUsdc:
          type: number
        expiresAt:
          type: string
          format: date-time
        expired:
          type: boolean
        signedAt:
          type: string
          format: date-time
        reason:
          type: string
        eip712:
          type: object
          properties:
            domain:
              type: object
              properties:
                name:
                  type: string
                version:
                  type: string
                chainId:
                  type: integer
            types:
              type: object
              additionalProperties: true
            primaryType:
              type: string
            allowedContract:
              type: string
      required:
        - active
    SubmitDelegationInput:
      type: object
      properties:
        delegator:
          type: string
        delegate:
          type: string
        allowedContract:
          type: string
        dailyLimitUsdc:
          type: number
        expiresAt:
          type: number
        signature:
          type: string
      required:
        - delegator
        - delegate
        - allowedContract
        - dailyLimitUsdc
        - expiresAt
        - signature
...
    RegistrationResult:
      type: object
      properties:
        success:
          type: boolean
        agentId:
          type: integer
        txHash:
          type: string
        error:
          type: string
      required:
        - success
...
    SwapInput:
      type: object
      properties:
        amount:
          type: number
          description: USDC amount to swap to ETH
      required:
        - amount
    SwapResult:
      type: object
      properties:
        success:
          type: boolean
        txHash:
          type: string
        amountIn:
          type: string
        amountOut:
          type: string
        error:
          type: string
        delegationDenied:
          type: boolean
        reason:
          type: string
      required:
        - success